[Ref] Assertions in macros tend to test for always-true conditions in certain cases (i.e. in the MPT extension writing macros). Selectively disable the relevant compiler warnings in the ASSERT macros.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5178 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_it.cpp
@@ -2115,7 +2115,7 @@ void CSoundFile::SaveExtendedSongProperties(FILE* f) const
 	if(!songArtist.empty())
 	{
 		std::string songArtistU8 = mpt::ToCharset(mpt::CharsetUTF8, songArtist);
-		uint32 length = mpt::saturate_cast<uint16>(songArtistU8.length());
+		uint16 length = mpt::saturate_cast<uint16>(songArtistU8.length());
 		WRITEMODULARHEADER(MAGIC4LE('A','U','T','H'), length);
 		mpt::IO::WriteRaw(f, songArtistU8.c_str(), length);
 	}
